# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ultra Computer API
  version: 0.9.0
  description: |
    Full REST API for Ultra Computer — an AI orchestration platform with multi-agent
    swarm coordination, autonomous scheduling, voice, browser automation, marketplace,
    identity, messaging, and protocol interoperability (A2A, MCP, CLI, webhooks).
servers:
  - url: http://localhost:5000
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer JWT token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    Conversation:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
        orchestratorModelId:
          type: string
          nullable: true
        createdAt:
          type: integer
        updatedAt:
          type: integer
...
paths:
...
  /api/conversations/{id}:
    get:
      operationId: getConversation
      summary: Get a conversation by ID
      tags: [Conversations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    patch:
      operationId: updateConversation
      summary: Update a conversation (title, status, orchestratorModelId)
      tags: [Conversations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                status:
                  type: string
                orchestratorModelId:
                  type: string
      responses:
        '200':
          description: Updated conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
    delete:
      operationId: deleteConversation
      summary: Delete a conversation and all its messages/tasks
      tags: [Conversations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /api/conversations/{id}/messages:
    get:
      operationId: getMessages
      summary: Get all messages for a conversation
      tags: [Conversations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
    post:
      operationId: sendMessage
      summary: Send a message and trigger orchestration
      tags: [Conversations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                role:
                  type: string
                  default: user
                modelId:
                  type: string
      responses:
        '200':
          description: Created message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
...
  /api/nip/sessions:
    get:
      operationId: listNIPSessions
      summary: List NIP sessions
      tags: [NIP]
      parameters:
        - name: state
          in: query
          schema:
            type: string
        - name: organizationId
          in: query
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: integer
        - name: to
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NIPSession'
    post:
      operationId: createNIPSession
      summary: Create a new NIP session
      tags: [NIP]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [instructorProfile, executorProfile]
              properties:
                instructorProfile:
                  type: object
                executorProfile:
                  type: object
                taskScope:
                  type: object
                accessTier:
                  type: string
                  enum: [public, verified, corporate, private]
      responses:
        '201':
          description: Created session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NIPSession'
...
  /api/nip/sessions/{id}:
    get:
      operationId: getNIPSession
      summary: Get a NIP session by ID
      tags: [NIP]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NIPSession'
        '404':
          description: Not found
  /api/nip/sessions/{id}/negotiate:
    post:
      operationId: negotiateNIPSession
      summary: Negotiate and activate a NIP session
      tags: [NIP]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Negotiated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NIPSession'
  /api/nip/sessions/{id}/messages:
    get:
      operationId: getNIPMessages
      summary: Get conversation for a NIP session
      tags: [NIP]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      operationId: sendNIPMessage
      summary: Send a message in a NIP session
      tags: [NIP]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role, type, content]
              properties:
                role:
                  type: string
                  enum: [instructor, executor, monitor, system]
                type:
                  type: string
                  enum: [instruction, feedback, question, acknowledgment, status_update, error_report, capability_query, capability_response, task_boundary, monitor_alert, human_override]
                content:
                  type: string
                  description: Max 50,000 chars
                metadata:
                  type: object
      responses:
        '201':
          description: Sent message
          content:
            application/json:
              schema:
                type: object
  /api/nip/sessions/{id}/pause:
    post:
      operationId: pauseNIPSession
      summary: Pause an active NIP session
      tags: [NIP]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Paused session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NIPSession'
  /api/nip/sessions/{id}/resume:
    post:
      operationId: resumeNIPSession
      summary: Resume a paused NIP session
      tags: [NIP]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Resumed session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NIPSession'
  /api/nip/sessions/{id}/terminate:
    post:
      operationId: terminateNIPSession
      summary: Force-terminate a NIP session
      tags: [NIP]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Terminated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NIPSession'
  /api/nip/sessions/{id}/complete:
    post:
      operationId: completeNIPSession
      summary: Mark a NIP session as successfully completed
      tags: [NIP]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Completed session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NIPSession'
  /api/nip/sessions/{id}/alerts:
    get:
      operationId: getNIPSessionAlerts
      summary: Get monitor alerts for a NIP session
      tags: [NIP]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
  /api/nip/sessions/{id}/report:
    get:
      operationId: getNIPReport
      summary: Get an existing report for a NIP session
      tags: [NIP]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Report or null
          content:
            application/json:
              schema:
                type: object
                nullable: true
    post:
      operationId: generateNIPReport
      summary: Generate a report for a NIP session
      tags: [NIP]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Generated report
          content:
            application/json:
              schema:
                type: object
  /api/nip/sessions/{id}/stream:
    get:
      operationId: streamNIPSession
      summary: Per-session SSE stream for NIP events
      tags: [NIP]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Server-Sent Events stream
          content:
            text/event-stream:
              schema:
                type: string
...